feat: organize WIP — MVP demo, Z-up fix, Isaac Sim tooling (#54)
* chore(config): ruff exclude .claude, add assets glob, fix per-file-ignores

- Exclude .claude/ from ruff linting
- Include *.usda in package data
- Restructure per-file-ignores as proper TOML table
- Add notebook E402 ignore


* fix(core): Z-up coordinate system, type annotations, import cleanup

- Migrate demo.py seabed/dock from Y-down to Z-down convention
- Add buoyancy force channel to _apply_forces kernel
- Integrate acoustics, currents, propulsion, vehicles into unified demo
- Expand integration.py with richer pipeline orchestration
- Fix fluid/wave.py coordinate handling
- Add type annotations (__init__.py, graph_capture.py)
- Minor import/lint fixes across fluid/, sensors/, propulsion/


* feat(training): expand Isaac Lab env with observation/reward refinements


* feat(mvp): underwater robot MVP demo + asset system + CLI command

- Add oceanscale/mvp.py: end-to-end underwater robot demo with
  BlueROV2, sonar, DVL, acoustics, currents, waypoint mission
- Add oceanscale/assets/: USD asset loader for BlueROV2 Heavy
- Add scripts/convert_bluerov2_gz_to_usd.py: Gazebo→USD converter
- Add CLI: oceanscale demo underwater-mvp [--steps N] [--output-json]


* feat(scripts): Isaac Sim demo + release verification tooling

- isaacsim_underwater_demo.py: standalone Isaac Sim underwater scene
- verify_isaacsim_underwater_demo.py: headless verification
- isaaclab_directrl_smoke.py: DirectRL env smoke test
- audit_underwater_robot_demo_launch.py: launch readiness audit
- verify_underwater_robot_demo_release.py: release gate checks


* fix(benchmarks): import cleanup and code quality across all benchmarks

Consistent imports, remove unused variables, fix type annotations
across 32 benchmark files.


* test: update existing tests + add MVP/assets/IsaacSim test coverage

- Update 18 existing tests to match API changes (Z-up, new imports)
- Add test_underwater_robot_mvp.py, test_assets.py
- Add test_bluerov2_gz_converter.py, test_isaacsim_underwater_demo.py
- Add release gate and launch audit tests


* fix(notebooks): update Colab notebooks and training scripts


* feat(website): add UnderwaterRobotDemo component + media assets

- Add UnderwaterRobotDemo.astro interactive component
- Add demo video/image/3D viewer assets
- Wire into index.astro (en + zh)


* docs: update HANDOFF, REFERENCES, AGENTS, marinegym extract


---------

+++ oceanscale/fluid/grid.py
@@ -670,12 +670,12 @@ def _sph_compute_density(
     i = wp.hash_grid_point_id(grid, tid)
     pi = pos[i]
     h2 = h * h
-    rho = float(0.0)  # noqa: UP018
+    rho = float(0.0)  # noqa: UP018 - Warp needs dynamic variables in kernels.
     h9 = h * h * h * h * h * h * h * h * h
     coeff = 315.0 / (64.0 * 3.14159265 * h9)
 
     query = wp.hash_grid_query(grid, pi, h)
-    j = int(0)
+    j = int(0)  # noqa: RUF046,UP018 - Warp hash query mutates this index.
     while wp.hash_grid_query_next(query, j):
         d = pos[j] - pi
         dist2 = d[0] * d[0] + d[1] * d[1] + d[2] * d[2]
@@ -712,16 +712,16 @@ def _sph_compute_forces(
     p_i = stiffness * (rho_i - rest_density)
 
     h2 = h * h
-    fx = float(0.0)  # noqa: UP018
-    fy = float(0.0)  # noqa: UP018
-    fz = float(0.0)  # noqa: UP018
+    fx = float(0.0)  # noqa: UP018 - Warp needs dynamic variables in kernels.
+    fy = float(0.0)  # noqa: UP018 - Warp needs dynamic variables in kernels.
+    fz = float(0.0)  # noqa: UP018 - Warp needs dynamic variables in kernels.
 
     h6 = h * h * h * h * h * h
     spiky_coeff = -45.0 / (3.14159265 * h6)
     visc_coeff = 45.0 / (3.14159265 * h6)
 
     query = wp.hash_grid_query(grid, pi, h)
-    j = int(0)
+    j = int(0)  # noqa: RUF046,UP018 - Warp hash query mutates this index.
     while wp.hash_grid_query_next(query, j):
         if i == j:
             continue
@@ -849,7 +849,7 @@ class SPHSolver:
         self.density = wp.zeros(n_particles, dtype=wp.float32, device=device)
         self.force = wp.zeros(n_particles, dtype=wp.vec3, device=device)
 
-        grid_dim = max(int(math.ceil(max(domain) / smoothing_length)), 4)
+        grid_dim = max(math.ceil(max(domain) / smoothing_length), 4)
         self.grid = wp.HashGrid(grid_dim, grid_dim, grid_dim, device=device)
 
     def build_spatial_hash(self):
